+++ pkg/solver.go
@@ -11,13 +11,13 @@ func tookSeconds(d time.Duration) string {
 	return fmt.Sprintf("%f seconds", d.Seconds())
 }
 
-func SolveWith(ctx context.Context, name string, solver func() int) {
+func SolveWith(ctx context.Context, name string, solver func() uint64) {
 	fmt.Fprintf(os.Stderr, "Hello From %s\n", name)
 	defer func(begin time.Time) {
         fmt.Fprintf(os.Stderr, "took: %s\n", tookSeconds(time.Since(begin)))
 	}(time.Now())
 
-	answerChannel := make(chan int, 1)
+	answerChannel := make(chan uint64, 1)
 	go func() {
         answerChannel <- solver()
 	}()
